chore: Rebuilt dfs1 with zfs support
The iso/autoinstall.nix file now also takes a flavor enum which lets you choose
a zfs or lvm layout.

The zfs layout is the only one that I know works. All the rest of the changes
are coming along for the ride cause I can't be arsed to split them into separate commits.