darwin: add restic setup for macos
Using git-transcrypt for now to store all the restic stuff as an sh script to be
sourced until something like sops or age runs on darwin/macos.

Next up is doing similar for nixos. Not sure if I should have restic on nixos
use sops or mirror darwin perfectly yet... That is a future me problem.