hm: Ok fine lets live with the stupid warning message
Looks like the flake version of nix behaves differently if ~/.nix-defexpr is a
symlink to /dev/null vs if it is a dir and has a channels symlink to /dev/null.

Who knew? I didn't, not sure how it worked on the home laptop. Probably a newer
nix difference who knows future me problem.