macos: conditionalize some home-manager setup
Adding a script force-paste to work around the nonsense that refuses to let you
paste. So fine, be that way, emulate a keyboard and pbpaste the clipboard and
type the text in that way.

Enabling the script in the script editor menu bar for now so its easy to (ab)use.