Try looping in the reopen logic
Stats is the biggest PITA at not staying open sometimes I swear. Might need to
brain on this more or let reopens happen in the background then check and just
reopen what seems to not be running again and have a circuit breaker for how
long to try with some backoff?d

Future me problem will see if this quick hacks good enough or not.