checkpoint, with nixos 23.11 getting a lot of infinite recursion
Not sure where/why in roles for now i'm working around it by just cheating but
it seems to be when I pass config in with the mkIf it recurses. Need to better
understand where that fits into attrvalue instatiation as it works with like
mutagen maybe its my hack attempts at using it in home-manager and nixos/darwin.