checkpoint: very simple setup
added an "installer" function to the zshrc for the iso installer setup.

The configuration.nix that gets built is wack af and doesn't output to the right
terminal for vm testing atm but whatever it installs.

I'll port over more stuff and get btrfs to do mirrors later. This is enough for now.