chore: Got more things working, vm-simple now autobuilds and works via deploy-rs
This is all a bit janky but its about where I want it to be. Need to port the
vm-mirror setup over, build a vm-raid setup with > 2 disks and I'm gtg.

The nix flake check is taking for freaking ever now though but its OK.

Also on the todo list is to figure out a better hostkey setup with ca's.

Everything in this commit for vm-simple is more of a test really.